requirements.txt
fastapi
uvicorn
networkx
numpy
pydantic
requests
streamlit
folium
streamlit-folium
python-dotenv